Fix issue where TaxCloud API is broken
We have seen issues with TaxCloud that have been reproducible for the last few hours, and have been intermittent since Saturday. The issue comes back with the following error in the traceback:

`suds.WebFault: Server raised fault: 'Server was unable to read request. ---> There is an error in XML document (1, 1083). ---> The string '' is not a valid Boolean value.'`

After some debugging, we have traced this to the `deliverBySender` having been sent as an empty string. This pull request aims to solve this problem for anyone else experiencing the issue.